:test_tube: refactor casts
