+++ cli/src/cast.rs
@@ -15,7 +15,7 @@ use ethers::{
     },
     providers::{Middleware, Provider},
     signers::{LocalWallet, Signer},
-    types::{Address, Chain, NameOrAddress, Signature, U256, U64},
+    types::{Address, Chain, NameOrAddress, Signature, U256},
     utils::get_contract_address,
 };
 use opts::{
@@ -36,9 +36,8 @@ use std::{
 use clap::{IntoApp, Parser};
 use clap_complete::generate;
 
-use crate::utils::read_secret;
+use crate::{cmd::Cmd, utils::read_secret};
 use eyre::WrapErr;
-use futures::join;
 
 #[tokio::main]
 async fn main() -> eyre::Result<()> {
@@ -542,64 +541,7 @@ async fn main() -> eyre::Result<()> {
             let selector = contract.abi().functions().last().unwrap().short_signature();
             println!("0x{}", hex::encode(selector));
         }
-        Subcommands::FindBlock { timestamp, rpc_url } => {
-            let ts_target = U256::from(timestamp);
-            let provider = Provider::try_from(rpc_url)?;
-            let last_block_num = provider.get_block_number().await?;
-            let cast_provider = Cast::new(provider);
-
-            let res = join!(cast_provider.timestamp(last_block_num), cast_provider.timestamp(1));
-            let ts_block_latest = res.0.unwrap();
-            let ts_block_1 = res.1.unwrap();
-
-            let block_num = if ts_block_latest.lt(&ts_target) {
-                // If the most recent block's timestamp is below the target, return it
-                last_block_num
-            } else if ts_block_1.gt(&ts_target) {
-                // If the target timestamp is below block 1's timestamp, return that
-                U64::from(1)
-            } else {
-                // Otherwise, find the block that is closest to the timestamp
-                let mut low_block = U64::from(1); // block 0 has a timestamp of 0: https://github.com/ethereum/go-ethereum/issues/17042#issuecomment-559414137
-                let mut high_block = last_block_num;
-                let mut matching_block: Option<U64> = None;
-                while high_block.gt(&low_block) && matching_block.is_none() {
-                    // Get timestamp of middle block (this approach approach to avoids overflow)
-                    let high_minus_low_over_2 = high_block
-                        .checked_sub(low_block)
-                        .ok_or_else(|| eyre::eyre!("unexpected underflow"))
-                        .unwrap()
-                        .checked_div(U64::from(2))
-                        .unwrap();
-                    let mid_block = high_block.checked_sub(high_minus_low_over_2).unwrap();
-                    let ts_mid_block = cast_provider.timestamp(mid_block).await?;
-
-                    // Check if we've found a match or should keep searching
-                    if ts_mid_block.eq(&ts_target) {
-                        matching_block = Some(mid_block)
-                    } else if high_block.checked_sub(low_block).unwrap().eq(&U64::from(1)) {
-                        // The target timestamp is in between these blocks. This rounds to the
-                        // highest block if timestamp is equidistant between blocks
-                        let res = join!(
-                            cast_provider.timestamp(high_block),
-                            cast_provider.timestamp(low_block)
-                        );
-                        let ts_high = res.0.unwrap();
-                        let ts_low = res.1.unwrap();
-                        let high_diff = ts_high.checked_sub(ts_target).unwrap();
-                        let low_diff = ts_target.checked_sub(ts_low).unwrap();
-                        let is_low = low_diff.lt(&high_diff);
-                        matching_block = if is_low { Some(low_block) } else { Some(high_block) }
-                    } else if ts_mid_block.lt(&ts_target) {
-                        low_block = mid_block;
-                    } else {
-                        high_block = mid_block;
-                    }
-                }
-                matching_block.unwrap_or(low_block)
-            };
-            println!("{}", block_num);
-        }
+        Subcommands::FindBlock(cmd) => cmd.run()?,
         Subcommands::Wallet { command } => match command {
             WalletSubcommands::New { path, password, unsafe_password } => {
                 let mut rng = thread_rng();
